requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.4.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.2=pypi_0
ca-certificates=2023.11.17=hbcca054_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
comm=0.2.1=pyhd8ed1ab_0
compressai=1.2.4=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
einops=0.7.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fonttools=4.47.0=pypi_0
fsspec=2023.12.2=pypi_0
gdown=4.7.1=pypi_0
idna=3.6=pypi_0
importlib-resources=6.1.1=pypi_0
ipykernel=6.28.0=pyhd33586a_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.0=py38h578d9bd_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.3=pypi_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
platformdirs=4.1.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch-msssim=1.0.0=pypi_0
pyzmq=25.1.0=py38h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
triton=2.1.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
